All tests now passing

+++ huffman.cpp
@@ -14,6 +14,7 @@
 #include <string>   // for STRING: binary representation of codes
 #include <sstream>
 #include <vector>
+#include <deque>
 
 #include "bnode.h"   // for BINARY_NODE class definition
 #include "vector.h"  // for VECTOR container
@@ -30,19 +31,34 @@ using namespace custom;
 
 class huffmanTree
 {
+public:
+   typedef custom::pair<std::string, float> PairType;
+   typedef BNode<PairType> NodeType;
 
 public:
-   typedef BNode<custom::pair<std::string, float>> NodeType;
    huffmanTree()
+       : pRoot(nullptr)
    {
    }
    huffmanTree(const std::string &value, float frequency)
    {
       pRoot = new NodeType(custom::pair<std::string, float>(value, frequency));
    }
+   huffmanTree(const PairType &pair)
+   {
+      pRoot = new NodeType(pair);
+   }
+   huffmanTree(NodeType *root)
+   {
+      pRoot = root;
+   }
    ~huffmanTree()
    {
-      deleteBTree(pRoot);
+      if (pRoot != nullptr)
+      {
+         deleteBTree(pRoot);
+         pRoot = nullptr;
+      }
    }
 
    void printCodes();
@@ -55,20 +71,73 @@ public:
          //not sure that's necessary yet.
          return pRoot->data.second;
       }
+      return 0;
+   }
+
+   void mergeLeft(huffmanTree *oldTree)
+   {
+      addLeft(pRoot, oldTree->pRoot);
+      //since we took ownership of the root
+      //we don't want destructor for huffmanTree
+      //to delete our child nodes. So set it to null.
+      oldTree->pRoot = nullptr;
+   }
+
+   void mergeRight(huffmanTree *oldTree)
+   {
+      addRight(pRoot, oldTree->pRoot);
+      //since we took ownership of the root
+      //we don't want destructor for huffmanTree
+      //to delete our child nodes. So set it to null.
+      oldTree->pRoot = nullptr;
+   }
+
+   std::string getHuffmanCode(std::string value)
+   {
+      std::string code;
+      getHuffmanCode(pRoot, code, value);
+      return code;
    }
 
 private:
    NodeType *pRoot;
+   //I'm SURE there's a better way to do this. seems way to complicated.
+   bool getHuffmanCode(NodeType *pNode, std::string &code, string value)
+   {
+      if (pNode == nullptr)
+         return false;
+
+      if (pNode->data.first == value)
+         return true; //found it!
+
+      if (pNode->pLeft != nullptr)
+      {
+         code.push_back('0');
+         if (getHuffmanCode(pNode->pLeft, code, value))
+            return true;
+         //get rid of the value we added
+         code.pop_back();
+      }
+
+      if (pNode->pRight != nullptr)
+      {
+         code.push_back('1');
+         if (getHuffmanCode(pNode->pRight, code, value))
+            return true;
+         code.pop_back();
+      }
+
+      return false;
+   }
 };
 
 void getTwoLowest(std::vector<huffmanTree *> trees, huffmanTree *&lowest, huffmanTree *&secondLowest)
 {
-
    for (std::vector<huffmanTree *>::iterator it = trees.begin(); it != trees.end(); ++it)
    {
       huffmanTree *p = *it;
 
-      if (secondLowest == nullptr || p->getValue() < secondLowest->getValue())
+      if (secondLowest == nullptr || (p->getValue() < secondLowest->getValue()))
       {
          if (lowest == nullptr || p->getValue() < lowest->getValue())
          {
@@ -98,6 +167,26 @@ void deleteTrees(std::vector<huffmanTree *> trees)
    }
 }
 
+std::vector<huffmanTree::PairType> readPairsFromFile(const string &fileName)
+{
+   std::ifstream infile(fileName);
+   std::string line;
+   std::vector<huffmanTree::PairType> pairs;
+
+   while (std::getline(infile, line))
+   {
+      std::istringstream iss(line);
+      std::string word;
+      float frequency;
+      if (!(iss >> word >> frequency))
+      {
+         break;
+      } // error
+      pairs.push_back(huffmanTree::PairType(word, frequency));
+   }
+   return pairs;
+}
+/*
 std::vector<huffmanTree *> readTreesFromFile(const string &fileName)
 {
    std::ifstream infile(fileName);
@@ -116,7 +205,9 @@ std::vector<huffmanTree *> readTreesFromFile(const string &fileName)
       huffmanTree *tree = new huffmanTree(word, frequency);
       trees.push_back(tree);
    }
+   return trees;
 }
+*/
 
 /*******************************************
  * HUFFMAN
@@ -124,26 +215,65 @@ std::vector<huffmanTree *> readTreesFromFile(const string &fileName)
  *******************************************/
 void huffman(const string &fileName)
 {
+   //get the pairs, rather than build trees immediately so we have
+   //the original values and order
+   std::vector<huffmanTree::PairType> pairs = readPairsFromFile(fileName);
 
-   std::vector<huffmanTree *> trees = readTreesFromFile(fileName);
+   //then convert the pairs into trees. I honestly wouldn't do this normally,
+   //because you could start with just the pairs. But the description of the
+   //assignment says to make a class to encapsulate the btree and to start
+   //with trees, then combine them until there is one. So we'll do that.
+   std::vector<huffmanTree *> trees;
+   for (std::vector<huffmanTree::PairType>::iterator it = pairs.begin(); it != pairs.end(); ++it)
+   {
+      trees.push_back(new huffmanTree(*it));
+   }
 
-   huffmanTree *pLowest;
-   huffmanTree *pSecondLowest;
+   //now loop through, getting the two lowest by value, and combine them until
+   //there's only one tree left
+   huffmanTree *pLowest = nullptr;
+   huffmanTree *pSecondLowest = nullptr;
 
    while (trees.size() > 1)
    {
       getTwoLowest(trees, pLowest, pSecondLowest);
 
-      //TODO: now we have to combine the two lowest into a new tree, and remove
-      //the two from "trees", and add the new one. (or collapse one into the other)
-      //this needs to happen until there is only one tree left in trees.
+      huffmanTree *pNewTree = new huffmanTree("", pLowest->getValue() + pSecondLowest->getValue());
+      pNewTree->mergeLeft(pLowest);
+      pNewTree->mergeRight(pSecondLowest);
 
-      //combining the trees might be tricky, because only "leaf" nodes are "real".
-      //probably the "huffmanTree.getValue" method isn't that useful here, because
-      //the value should be calculated from the leaves under it?
+      //remove the two old trees from the list
+      std::vector<huffmanTree *>::iterator it = trees.begin();
+      while (it != trees.end())
+      {
+         if ((pLowest != nullptr) && (*it == pLowest))
+         {
+            //replace the lowest with the new tree
+            *it = pNewTree;
+            delete pLowest;
+            pLowest = nullptr;
+         }
+         else if ((pSecondLowest != nullptr) && (*it == pSecondLowest))
+         {
+            //erase this one.
+            it = trees.erase(it);
+            delete pSecondLowest;
+            pSecondLowest = nullptr;
+         }
+         else
+         {
+            ++it;
+         }
+      }
    }
 
+   huffmanTree *finalTree = trees[0];
+
    //TODO: print the results
+   for (std::vector<huffmanTree::PairType>::iterator it = pairs.begin(); it != pairs.end(); ++it)
+   {
+      std::cout << it->first << " = " << finalTree->getHuffmanCode(it->first) << std::endl;
+   }
 
    deleteTrees(trees);
 }
